creating seperate files for cg testcases
