+++ tests/functional-tests/pytest.ini
@@ -23,3 +23,4 @@ markers =
     remotecluster: remote cluster notation
     localcluster: local cluster notation
     csioperator: csi operator testcases
+    cg: consistency group testcases
